Enhancement: 事务、分页、查询优化

+++ infra-market-server-vertx/src/main/kotlin/io/infra/market/vertx/dto/CommonDto.kt
@@ -21,8 +21,8 @@ data class ApiData<T>(
 data class PageResultDto<T>(
     val records: List<T>,
     val total: Long,
-    val page: Long,
-    val size: Long
+    val page: Int,
+    val size: Int
 )
 
 /**
